1048. Longest String Chain
