+++ src/main/kotlin/me/bossm0n5t3r/leetcode/longestStringChain/README.md
