+++ src/test/kotlin/me/bossm0n5t3r/leetcode/longestStringChain/LongestStringChainTest.kt
@@ -1,10 +1,10 @@
-package longestStringChain
+package me.bossm0n5t3r.leetcode.longestStringChain
 
 import org.junit.jupiter.api.Test
 import kotlin.test.assertEquals
 
 internal class LongestStringChainTest {
-    private val longestStringChain = LongestStringChain()
+    private val longestStringChain = LongestStringChain.Solution()
 
     private data class LongestStringChainTestData(
         val words: Array<String>,
